username validation for delete and update method ok

+++ routes/ideas.js
@@ -66,33 +66,39 @@ router.post('/', async (req, res) => {
 const mongoose = require('mongoose');
 
 router.put('/:id', async (req, res) => {
- const ideaId = req.params.id;
-    const { text, tag, username } = req.body;
-    const updateFields = {};
-
-    if (!isValidObjectId(ideaId)) {
-        return res.status(400).json({ succeed: false, error: 'Invalid Idea ID' });
-    }
-    
-    if (text) updateFields.text = text;
-    if (tag) updateFields.tag = tag;
-    if (username) updateFields.username = username;
-
-    if (Object.keys(updateFields).length === 0) {
-        return res.status(400).json({ succeed: false, error: 'No fields to update' });
-    }
 
     try {
-        const updatedIdea = await Idea.findOneAndUpdate(
-            { _id: ideaId },
-            { $set: updateFields },
-            { new: true }
-        );
-        if (updatedIdea === null) {
-            return res.status(404).json({ succeed: false, error: 'Idea not found' });
-        }
+        const ideaId = req.params.id;
+        const idea=await Idea.findById(ideaId);
+        if(idea.username===req.body.username){
 
-        res.status(200).json({ succeed: true, data: updatedIdea });
+            const { text, tag, username } = req.body;
+            const updateFields = {};
+        
+            if (!isValidObjectId(ideaId)) {
+                return res.status(400).json({ succeed: false, error: 'Invalid Idea ID' });
+            }
+            if (text) updateFields.text = text;
+            if (tag) updateFields.tag = tag;
+            if (username) updateFields.username = username;
+        
+            if (Object.keys(updateFields).length === 0) {
+                return res.status(400).json({ succeed: false, error: 'No fields to update' });
+            }
+            const updatedIdea = await Idea.findOneAndUpdate(
+                { _id: ideaId },
+                { $set: updateFields },
+                { new: true }
+            );
+            if (updatedIdea === null) {
+                return res.status(404).json({ succeed: false, error: 'Idea not found' });
+            }
+            res.status(200).json({ succeed: true, data: updatedIdea });
+        }else{
+            // Username don't match
+            res.status(403).json({ succeed: false, error: 'you are not autorized to delete this ressource' });
+        }
+       
     } catch (error) {
         res.status(500).json({ succeed: false, error: error.message });
     }
@@ -107,16 +113,22 @@ router.put('/:id', async (req, res) => {
  */
 router.delete('/:id', async (req, res) => {
     try {
-        const ideaId = req.params.id;
-        if (!isValidObjectId(ideaId)) {
-            return res.status(400).json({ succeed: false, error: 'Invalid idea ID' });
+        const idea=await Idea.findById(req.params.id);
+        if(idea.username===req.body.username){
+            const ideaId = req.params.id;
+            if (!isValidObjectId(ideaId)) {
+                return res.status(400).json({ succeed: false, error: 'Invalid idea ID' });
+            }
+            const deletedIdea = await Idea.findOneAndDelete({ _id: ideaId });
+            if (!deletedIdea) {
+                return res.status(404).json({ succeed: false, error: 'Idea not found' });
+            }
+            res.status(200).json({ succeed: true, data: deletedIdea });
+        }else{
+            res.status(403).json({ succeed: false, error: 'you are not autorized to delete this ressource' });
         }
-        const deletedIdea = await Idea.findOneAndDelete({ _id: ideaId });
-        if (!deletedIdea) {
-            return res.status(404).json({ succeed: false, error: 'Idea not found' });
-        }
-        res.status(200).json({ succeed: true, data: deletedIdea });
     } catch (error) {
+        console.log(error);
         res.status(500).json({ succeed: false, error: error.message });
     }
 });
